requirements.txt
streamlit
langchain
langchain-community
langchain-openai
faiss-cpu
openai
python-dotenv
tiktoken
pypdf
